Allows deletion of object without refresh
